+++ packages/databricks-vscode/src/cli/CliWrapper.ts
@@ -194,7 +194,10 @@ export class CliWrapper {
         try {
             res = await execFile(cmd.command, cmd.args, {
                 env: {
-                    ...EnvVarGenerators.getEnvVarsForCli(configfilePath),
+                    ...EnvVarGenerators.getEnvVarsForCli(
+                        this.extensionContext,
+                        configfilePath
+                    ),
                     ...EnvVarGenerators.getProxyEnvVars(),
                 },
             });
@@ -295,7 +298,10 @@ export class CliWrapper {
             const {stdout, stderr} = await execFile(cmd[0], cmd.slice(1), {
                 cwd: workspaceFolder.fsPath,
                 env: {
-                    ...EnvVarGenerators.getEnvVarsForCli(configfilePath),
+                    ...EnvVarGenerators.getEnvVarsForCli(
+                        this.extensionContext,
+                        configfilePath
+                    ),
                     ...EnvVarGenerators.getProxyEnvVars(),
                     ...authProvider.toEnv(),
                     ...this.getLogginEnvVars(),
@@ -328,7 +334,17 @@ export class CliWrapper {
         configfilePath?: string,
         logger?: logging.NamedLogger
     ) {
-        const cmd = [this.cliPath, "bundle", "summary", "--target", target];
+        const cmd = [
+            this.cliPath,
+            "bundle",
+            "summary",
+            "--target",
+            target,
+            // Forces the CLI to regenerate local terraform state and pull the remote state.
+            // Regenerating terraform state is useful when we want to ensure that the provider version
+            // used in the local state matches the bundled version we supply with the extension.
+            "--force-pull",
+        ];
 
         logger?.info(
             `Refreshing bundle configuration for target ${target}...`,
@@ -340,7 +356,10 @@ export class CliWrapper {
             const {stdout, stderr} = await execFile(cmd[0], cmd.slice(1), {
                 cwd: workspaceFolder.fsPath,
                 env: {
-                    ...EnvVarGenerators.getEnvVarsForCli(configfilePath),
+                    ...EnvVarGenerators.getEnvVarsForCli(
+                        this.extensionContext,
+                        configfilePath
+                    ),
                     ...EnvVarGenerators.getProxyEnvVars(),
                     ...authProvider.toEnv(),
                     ...this.getLogginEnvVars(),
@@ -369,6 +388,7 @@ export class CliWrapper {
     getBundleInitEnvVars(authProvider: AuthProvider) {
         return removeUndefinedKeys({
             ...EnvVarGenerators.getEnvVarsForCli(
+                this.extensionContext,
                 workspaceConfigs.databrickscfgLocation
             ),
             ...EnvVarGenerators.getProxyEnvVars(),
@@ -425,7 +445,10 @@ export class CliWrapper {
 
         // Add python executable to PATH
         const executable = await pythonExtension.getPythonExecutable();
-        const cliEnvVars = EnvVarGenerators.getEnvVarsForCli(configfilePath);
+        const cliEnvVars = EnvVarGenerators.getEnvVarsForCli(
+            this.extensionContext,
+            configfilePath
+        );
         let shellPath = cliEnvVars.PATH;
         if (executable) {
             shellPath = `${path.dirname(executable)}${
@@ -491,7 +514,10 @@ export class CliWrapper {
         options: SpawnOptionsWithoutStdio;
     } {
         const env: Record<string, string> = removeUndefinedKeys({
-            ...EnvVarGenerators.getEnvVarsForCli(configfilePath),
+            ...EnvVarGenerators.getEnvVarsForCli(
+                this.extensionContext,
+                configfilePath
+            ),
             ...EnvVarGenerators.getProxyEnvVars(),
             ...authProvider.toEnv(),
             ...this.getLogginEnvVars(),