+++ packages/databricks-vscode-types/CHANGELOG.md
@@ -1,3 +1,15 @@
+# Release: v2.0.4-preview
+
+## packages/databricks-vscode-types
+
+## (2024-04-09)
+
+# Release: v2.0.3-preview
+
+## packages/databricks-vscode-types
+
+## (2024-04-08)
+
 # Release: v2.0.2-preview
 
 ## packages/databricks-vscode-types